requirements.txt
azure-ai-formrecognizer==3.3.3
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.35.0
azure-storage-blob==12.26.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.3
cryptography==45.0.5
cycler==0.12.1
fonttools==4.59.0
idna==3.10
isodate==0.7.2
kiwisolver==1.4.8
matplotlib==3.10.5
msrest==0.7.1
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pycparser==2.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
seaborn==0.13.2
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
